fix(test): rename fluent.ts to fluent.test.ts
